Better suggestion ordering
This should improve the relevancy of the tag suggestions; the earlier in the tag label the query substring occurs, the higher up the list it will appear. The example here was a search for 'prog', which previously brought back 'metaprogramming' before 'programming' (because 'm' is before 'p'...), and for that example we now get the correct result order.